+++ src/feno/debug.py
@@ -1,0 +1,6 @@
+class Db:
+    enable = False
+    @staticmethod
+    def print(*args, **kwargs):
+        if Db.enable:
+            print(*args, **kwargs)