+++ src/feno/decoder.py
@@ -1,6 +1,6 @@
 import chardet
-import os
-
+from .debug import Db
+from .text import Text
 class Decoder:
 
     @staticmethod
@@ -8,13 +8,14 @@ class Decoder:
         with open(file_path, "rb") as file:
             raw_data = file.read()
             enc_dict = chardet.detect(raw_data)
+            Db.print(Text("{g} encoding: {y}", file_path, str(enc_dict)))
             encoding = enc_dict["encoding"]
-            if encoding is None:
+            if encoding is None or enc_dict["confidence"] < 0.95 :
                 encoding = "utf-8"
             try:
                 content = raw_data.decode(encoding)
             except UnicodeDecodeError as e:
-                content = raw_data.decode("utf-8")
+                content = raw_data.decode(enc_dict["encoding"])
             return content
 
     @staticmethod